Added new solutions

+++ new_solution.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-YEAR=2015
-DAY=24
+YEAR=2021
+DAY=21
 
-touch $YEAR/$DAY.txt
-cp _template.py $YEAR/$DAY.py
+
+if [ ! -f $YEAR/$DAY.txt ]; then
+    touch $YEAR/$DAY.txt
+    cp _template.py $YEAR/$DAY.py
+fi
